collector-agent/requirements.txt
gevent==1.4.0
greenlet==0.4.15
grpcio==1.26.0
grpcio-tools==1.26.0
protobuf==3.9.1
psutil==5.6.3
six==1.12.0
thrift==0.11.0
pyzmq==18.1.1

